Add working draft of invertebrate summary table
Using the (now deprecated) Standard Expressions of the tidyverse create a summary table of the data grouped by taxon that gives information on the total abdunance of the taxon, the number of sites visited that contained the taxon, and additional summary information for the taxon.

The working draft was created with test data and so the input file path and the output file path must be changed for the actual data.