Add start of clean version of Combine_InvLabData.R
The goal is to have a version of Combine_InvLabData.R that acts
as a sourceable script. This is the start of it which will have
additional work added. Ideally this would just have been done on
Combine_InvLabData.R, but this work can simply be _moved_ to it later.